add a country info fragment
