+++ app/src/main/res/layout/fragment_country_intro.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".countryContacts.CountryInfoFragment">
+    tools:context=".countryContacts.CountryIntroFragment">
 
     <ImageView
         android:id="@+id/fish_image"
@@ -15,20 +15,21 @@
         app:layout_constraintTop_toTopOf="parent"/>
     <TextView
         android:id="@+id/title"
-        android:text="Write a letter to:"
+        android:text="@string/write_a_letter_and_spread_the_word"
         android:textColor="@android:color/black"
+        android:textAlignment="center"
         android:textSize="20dp"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingTop="20dp"
-        android:paddingLeft="20dp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/fish_image"/>
     <TextView
-        android:id="@+id/adress_text"
-        android:text="Justin Trudeau\nHouse of Commons\nOttawa, ON  K1A 0A6"
+        android:id="@+id/additional_text"
+        android:text="@string/write_a_letter_desctiprion"
         android:padding="20dp"
+        android:textAlignment="center"
         android:textSize="16dp"
         android:textColor="@color/dark_gray"
         android:layout_width="match_parent"
@@ -36,15 +37,15 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/title"/>
-    <TextView
-        android:text="Stamp, mail it off, and then enter it in the Fatechanger Map’s Tracker. "
-        android:textColor="@android:color/black"
-        android:padding="20dp"
-        android:textSize="18dp"
-        android:layout_width="match_parent"
+    <Button
+        android:id="@+id/write_to_where_button"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:text="@string/write_to_where"
+        style="@style/small_button_style"
+        android:layout_marginBottom="20dp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/adress_text"/>
+        app:layout_constraintBottom_toBottomOf="parent"/>
 
 </android.support.constraint.ConstraintLayout>
