+++ app/src/main/res/values/colors.xml
@@ -5,6 +5,7 @@
     <color name="colorAccent">#20c2de</color>
     <color name="lightgray">#E5E5E5</color>
     <color name="gray">#b2b2b2</color>
+    <color name="dark_gray">#4a4a4a</color>
     <color name="gold">#DAA520</color>
     <color name="greenish_blue">#2672b3</color>
     <color name="blueish_subtitle_color">#4a81b6</color>
